Add an ``backwards`` option to the autopaginate tag.
Optionally, ``backwards`` may be passed as the last argument to the
template tag. Useful when the ``object_list`` is in a reverse chronological
order and you'd like to start paginating from the opposite direction.
Meaning, if there are 6 pages of results, visiting the item for the first
time will start with the 6th page of results instead of the 1st.

{% autopaginate object_list %}

{% autopaginate object_list 10 backwards %}

Note it is still up to you to reverse the ``object_list`` before using
the ``backwards`` option.